Updated the structure of the test directory
Previously, the test needed to be called from a terminal. It would not
work if called directly from RStudio. I`ve updated the tests methods so
they work from RStudio. I think this will be helpfull when submitting
the code for CRAN as it will be able to run tests automatically (without
an additional support script). I will need to write a new test.R script
but I will do that later.